% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countValues.R
\name{countValues}
\alias{countValues}
\title{Count Occurrences of Values in Specified Columns}
\usage{
countValues(data, group_by, na.rm = FALSE)
}
\arguments{
\item{data}{A data frame containing the data to be grouped and counted.}

\item{group_by}{A character vector of column names to group by.}

\item{na.rm}{Logical indicating whether to remove rows with NA values in the grouping columns.}
}
\value{
A data frame with one row per unique combination of values in the specified columns, with a \code{count} column indicating the occurrences.
}
\description{
\code{countValues} is a function that groups data by specified columns and counts the number of occurrences in each group.
}
\examples{
library(dplyr)
test_data <- data.frame(
  species_name = c("SERRULATA", "CERASIFERA", "PLATANOIDES", "RUBRUM", NA, "AMERICANA"),
  neighbourhood = c("Downtown", "West End", "Downtown", "Kitsilano", "West End", "Kitsilano"),
  diameter = c(10, 20, 30, 40, NA, 15)
)
countValues(test_data, group_by = c("species_name", "neighbourhood"), na.rm = TRUE)
}
